+++ packages/xcm-cfg/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @galacticcouncil/xcm-cfg
 
+## 1.2.0
+
+### Minor Changes
+
+- unique support
+- nodle support
+
+## 1.1.1
+
+### Patch Changes
+
+- Fix bifrost to hydradx xcm, extrinsic account support
+
 ## 1.1.0
 
 ### Minor Changes